+++ tests/sdk/testTransfer.py
@@ -30,6 +30,7 @@ class TestTransferGet(TestCase):
             before=date.today(),
             status="failed",
             tags=["iron", "bank"],
+            tax_id="012.345.678-90",
             transaction_ids=["1", "2", "3"],
         )
         self.assertEqual(len(list(transfers)), 0)